Push changes to beta branch

+++ Dev/QConnect/src/screens/TeacherScreens/ClassTabs/ClassEditScreen.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { ScrollView, View, StyleSheet, TextInput, FlatList, TouchableWithoutFeedback, Keyboard } from "react-native";
+import { ScrollView, View, StyleSheet, TextInput, FlatList, TouchableWithoutFeedback, Keyboard, Text, Alert } from "react-native";
 import Toast, { DURATION } from 'react-native-easy-toast'
 import { connect } from "react-redux";
 import StudentCard from "components/StudentCard";
@@ -9,7 +9,9 @@ import { deleteStudent } from "model/actions/deleteStudent";
 import { addStudent } from "model/actions/addStudent";
 import QcActionButton from "components/QcActionButton";
 import studentImages from "config/studentImages";
-import strings from "../../../../config/strings";
+import { Icon } from 'react-native-elements';
+import strings from "config/strings";
+import mapStateToCurrentClassProps from "screens/TeacherScreens/helpers/mapStateToCurrentClassProps";
 
 export class ClassEditScreen extends Component {
 
@@ -59,8 +61,7 @@ export class ClassEditScreen extends Component {
 
   //Tests whether the entered input is already a student that exists in the given class
   studentNameExists() {
-    const { params } = this.props.navigation.state;
-    const students = this.props.classes[params && params.classIndex ? params.classIndex : 0].students;
+    const students = this.props.students;
     const input = this.state.newStudentName.trim();
 
     //Will search if there is a student with the same name or not, if there is, it will return the 
@@ -124,10 +125,7 @@ export class ClassEditScreen extends Component {
   // ------- Render method: Main entry to render the screen's UI ------------------
 
   render() {
-    const { params } = this.props.navigation.state;
-    const { deleteStudent, addStudent, classes } = this.props;
-
-    classIndex = params && params.classIndex ? params.classIndex : 0;
+    const { deleteStudent, addStudent, classIndex, students } = this.props;
 
     if (this.state.highlightedImagesIndices.length == 0) {
       return false;
@@ -145,10 +143,12 @@ export class ClassEditScreen extends Component {
           />
 
           <View ID={classIndex} style={styles.inputContainer}>
+            <Text style={styles.inputName}>{strings.EnterYourStudentsName}</Text>
             <TextInput
-              placeholder={strings.EnterNewStudentsName}
+              placeholder={strings.StudentName}
               onChangeText={newStudentName => this.setState({ newStudentName })}
               value={this.state.newStudentName}
+              style={{ paddingBottom: 10 }}
             />
             <ImageSelectionRow
               images={studentImages.images}
@@ -163,7 +163,7 @@ export class ClassEditScreen extends Component {
             />
           </View>
           <FlatList
-            data={classes[classIndex].students}
+            data={students}
             keyExtractor={(item, index) => item.name} // fix, should be item.id (add id to classes)
             renderItem={({ item, index }) => (
               <StudentCard
@@ -171,12 +171,31 @@ export class ClassEditScreen extends Component {
                 studentName={item.name}
                 profilePic={studentImages.images[item.imageId]}
                 background={colors.white}
-                onPress={() =>
-                  deleteStudent(
-                    classIndex,
-                    index
-                  )
-                }
+                onPress={() => { }}
+                comp={<Icon
+                  name='user-times'
+                  size={25}
+                  type='font-awesome'
+                  color={colors.primaryLight}
+                  onPress={() => {
+                    Alert.alert(
+                      'Delete Student',
+                      'Are you sure you want to delete this student?',
+                      [
+                        {
+                          text: 'Delete', onPress: () => {
+                            deleteStudent(
+                              classIndex,
+                              index
+                            );
+                          }
+                        },
+
+                        { text: 'Cancel', style: 'cancel' },
+                      ]
+                    );
+                  }}
+                />}
               />
             )}
           />
@@ -204,12 +223,16 @@ const styles = StyleSheet.create({
   classTitle: {
     color: colors.primaryDark,
     fontSize: 25
+  },
+  inputName: {
+    fontFamily: "regular",
+    fontSize: 18,
+    paddingBottom: 10
   }
 });
 
-const mapStateToProps = state => {
-  const { classes } = state.data.teachers[0];
-  return { classes };
+const mapStateToProps = (state) => {
+  return mapStateToCurrentClassProps(state)
 };
 
 const mapDispatchToProps = dispatch =>
